jsonrcp: * added WebSocketMessageReader. * refactored code to create different classes for all the different readers. * add sub call to init the socket.io web socket.
jsonrpc module:
* compiles fine with the changes in place.
