+++ jsonrpc/src/message-readers/data-callback.ts
@@ -1,0 +1,5 @@
+import { Message } from '../messages';
+
+export interface DataCallback {
+	(data: Message): void;
+}
