+++ jsonrpc/src/message-readers/message-info.ts
@@ -1,0 +1,4 @@
+export interface PartialMessageInfo {
+	messageToken: number;
+	waitingTime: number;
+}
